Fix castid index issue
Previously , the cast id and the table does not match. The cast id has length 47734 and the length for beg is 29793. This is because there are sections in the original file names "END OF BIOLOGY SECTION". To be precise, there are tables where cast id is assigned, but no environmental data is taken, just biology data (include some catch records). 
To address this problem, cast id corresponding to tables with environmental records are tracked and assign the corresponding table.